Small tag manager redesign, refactor subprojects
Deleted 'customext' subproject as it is not needed.

Refactored the 'tags' subproject by adding the files to the 'riscv'
subproject to avoid unnecessary dependency management.

Redesign of tag manager:
- Added tag_memory_t class, that represents the main memory used by
the tag coprocessor. It is meant to tag the data and to hold the
policy. The object is a part of the simulator class.
- Refactored tag_manager_t class to include only register
tags (including PC). It is a part of the processor class. The class
holds a pointer to the tag_memory_t object, that is shared between all
the processors.