First try at tag propagation, add some instructions
The basic instruction set and the M instruction have tag propagation
code added.

The tag propagation works by having a global tag_memory_t object,
which holds the policy and the tag memory. It implements simif_t
interface, so it can read and write into tag memory. The tag memory is
constructed by duplicating the bus from sim_t and all abstract devices
that are added during the setup phase of the program.

The loading of tagged memory is added by overriding the PREAD syscall,
where we read from a 'tag' ELF file the same bytes as are read from
the PREAD target.

Each processor has a tag_manager_t object attached to it. The object
handles tag propagation by holding the register tag data and queries
the tag_memory_t for policy and tag memory information. It has its own
MMU, which accesses the memory from the tag_memory_t object (that is
why tag_memory_t implements simif_t, so that MMU can use it).

Problems:
- Some loads and stores throw traps, which are supposed to be handled
by trap handlers. This becomes a problem for the MMU that is used with
the tag_manager_t, as we cannot achieve the same result.