Fix MMU to fetch / load / store tags and data at the same time
The MMU now takes and returns an std::pair<data, tag>. It takes a
pointer to the tag_memory_t object so it can access the tag memory.

The TLB is expanded so it holds tag information. Fetch also returns PC
tag. The execute functions for instructions take pc tag as an
argument.

The tag manager propagate functions are redesigned to align with the
changes. The load and store functions are templates. The load
function combines the tag of the loaded byte / half-word / word into
one tag together with register information. The store function
computes the tag that will be stored into the tag memory.

TODO check branch and jump propagation rules.