Add optional tag checking of jump addresses
Branch and jump propagation functions now support jump address tag
check, if we choose to do it. It can be enabled with

```
$ ./conifgure --enable-tag-jump-check
```

The tag manager uses the processor's MMU to fetch the instruction on
the given address. This leads to more memory traps, but it dosen't
seem to cause problems.