Add tag memory to HTIF loads / stores
The HTIF interface had defined it's own store and load functions that
are used with syscalls.

We added additional parameters for tag buffers to the functions and
refactored the code to work.

There are a lot of loose ends with the function calls. In cases where
we don't know that we need the tag memory for specific calls, the
function takes 'unknown' tags.

Some classes don't do anything with the tag memory as we don't know
their intended use.

Refactored perimiter guards and syscalls to work with the new load /
store functions.